Функции вычисления перенесены в модуль calculations.py
